backend/requirements.txt
asgiref==3.5.0
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
eventlet==0.33.0
greenlet==1.1.2
idna==3.3
mysqlclient==2.1.0
PyJWT==2.3.0
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
websocket-client==1.3.2
